gcloud/requirements.txt
requests
pandas
matplotlib
pycountry
google-cloud-storage
